Editor: Move Property and Scintilla files into folders
