+++ Libraries/Editor/EditorStandard.hpp
@@ -46,19 +46,19 @@ public:
 #include "PropertyInterface.hpp"
 #include "UndoPropertyInterface.hpp"
 #include "MultiPropertyInterface.hpp"
-#include "PropertyWidget.hpp"
-#include "PropertyWidgets.hpp"
-#include "PropertyWidgetObject.hpp"
-#include "BasicPropertyEditors.hpp"
-#include "PropertyView.hpp"
-#include "VariantEditor.hpp"
-#include "DataBaseTemplate.hpp"
-#include "GraphView.hpp"
+#include "Property/PropertyWidget.hpp"
+#include "Property/PropertyWidgets.hpp"
+#include "Property/PropertyWidgetObject.hpp"
+#include "Property/BasicPropertyEditors.hpp"
+#include "Property/PropertyView.hpp"
+#include "Property/VariantEditor.hpp"
+#include "Property/DataBaseTemplate.hpp"
+#include "Property/GraphView.hpp"
 #include "Editor/TileLayout.hpp"
 #include "PreviewWidget.hpp"
-#include "TileView.hpp"
+#include "Property/TileView.hpp"
 #include "TreeViewFilter.hpp"
-#include "TreeView.hpp"
+#include "Property/TreeView.hpp"
 #include "TreeViewSearch.hpp"
 #include "ItemList.hpp"
 #include "ListView.hpp"
