+++ Libraries/Editor/Scintilla/ScintillaPlatformZero.cpp
