+++ Libraries/Editor/Scintilla/ScintillaPlatformZero.hpp
